Removed managed closures from iomgr.
The only remaining need for those (whereby the closure frees itself upon
completion) remains in security/credentials.c, where the code has been
inlined (to be removed at a later time).
